# sic_backend_mvp_jwt_sqlite/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DueSpark API",
...
    "version": "1.0.0",
    "contact": {
      "name": "DueSpark Support",
      "email": "support@duespark.com",
      "url": "https://duespark.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
...
    "/api/client-portal/access": {
      "post": {
        "tags": [
          "client-portal"
        ],
        "summary": "Request Client Access",
        "description": "Generate access token for client portal (public endpoint)",
        "operationId": "request_client_access_api_client_portal_access_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientAccessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/client-portal/portal/{access_token}": {
      "get": {
        "tags": [
          "client-portal"
        ],
        "summary": "Get Client Portal",
        "description": "Get client portal dashboard (public endpoint)",
        "operationId": "get_client_portal_api_client_portal_portal__access_token__get",
        "parameters": [
          {
            "name": "access_token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Access Token"
            }
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email",
              "description": "Client email for verification",
              "title": "Email"
            },
            "description": "Client email for verification"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/client-portal/invoice/{invoice_id}/payment": {
      "get": {
        "tags": [
          "client-portal"
        ],
        "summary": "Get Invoice Payment Page",
        "description": "Get branded payment page for specific invoice (public endpoint)",
        "operationId": "get_invoice_payment_page_api_client_portal_invoice__invoice_id__payment_get",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invoice Id"
            }
          },
          {
            "name": "access_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Client access token",
              "title": "Access Token"
            },
            "description": "Client access token"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/client-portal/invoice/{invoice_id}/pay": {
      "post": {
        "tags": [
          "client-portal"
        ],
        "summary": "Process Invoice Payment",
        "description": "Process payment for invoice (public endpoint)",
        "operationId": "process_invoice_payment_api_client_portal_invoice__invoice_id__pay_post",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invoice Id"
            }
          },
          {
            "name": "access_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Client access token",
              "title": "Access Token"
            },
            "description": "Client access token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/client-portal/message": {
      "post": {
        "tags": [
          "client-portal"
        ],
        "summary": "Send Client Message",
        "description": "Send message from client to vendor (public endpoint)",
        "operationId": "send_client_message_api_client_portal_message_post",
        "parameters": [
          {
            "name": "access_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Client access token",
              "title": "Access Token"
            },
            "description": "Client access token"
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email",
              "description": "Client email",
              "title": "Email"
            },
            "description": "Client email"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/client-portal/communication/{access_token}": {
      "get": {
        "tags": [
          "client-portal"
        ],
        "summary": "Get Communication History",
        "description": "Get communication history between client and vendor (public endpoint)",
        "operationId": "get_communication_history_api_client_portal_communication__access_token__get",
        "parameters": [
          {
            "name": "access_token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Access Token"
            }
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email",
              "description": "Client email",
              "title": "Email"
            },
            "description": "Client email"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/client-portal/feedback": {
      "post": {
        "tags": [
          "client-portal"
        ],
        "summary": "Submit Client Feedback",
        "description": "Submit client feedback and testimonial (public endpoint)",
        "operationId": "submit_client_feedback_api_client_portal_feedback_post",
        "parameters": [
          {
            "name": "access_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Client access token",
              "title": "Access Token"
            },
            "description": "Client access token"
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email",
              "description": "Client email",
              "title": "Email"
            },
            "description": "Client email"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/client-portal/receipt/{payment_id}": {
      "get": {
        "tags": [
          "client-portal"
        ],
        "summary": "Get Payment Receipt",
        "description": "Get payment receipt (public endpoint)",
        "operationId": "get_payment_receipt_api_client_portal_receipt__payment_id__get",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/client-portal/demo": {
      "get": {
        "tags": [
          "client-portal"
        ],
        "summary": "Get Demo Portal",
        "description": "Get demo client portal for showcasing (public endpoint)",
        "operationId": "get_demo_portal_api_client_portal_demo_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClientAccessRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "invoice_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Number"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "ClientAccessRequest"
      },
...
      "ClientFeedbackRequest": {
        "properties": {
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "feedback": {
            "type": "string",
            "title": "Feedback"
          },
          "public_testimonial": {
            "type": "boolean",
            "title": "Public Testimonial",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "rating",
          "feedback"
        ],
        "title": "ClientFeedbackRequest"
      },
      "ClientMessageRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "invoice_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invoice Id"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ClientMessageRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentRequest": {
        "properties": {
          "payment_method": {
            "type": "string",
            "title": "Payment Method"
          },
          "card_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Card Token"
          },
          "billing_address": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Address"
          }
        },
        "type": "object",
        "required": [
          "payment_method"
        ],
        "title": "PaymentRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from `/auth/login` endpoint"
      }
    }
  },
...
  "servers": [
    {
      "url": "https://api.duespark.com",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.duespark.com",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ]
}